Avoid names collisions using the standard FreeCAD external workbenches structure
