+++ freecad/invgears/newInternalCmd.py
@@ -1,7 +1,7 @@
 # ***************************************************************************
 # *   Copyright (c) 2021 Sebastian Ernesto García <sebasg@outlook.com>      *
 # *                                                                         *
-# *   newSlaveMasterCmd.py                                                  *
+# *   newInternalCmd.py                                                     *
 # *                                                                         *
 # *   This program is free software; you can redistribute it and/or modify  *
 # *   it under the terms of the GNU Lesser General Public License (LGPL)    *
@@ -26,66 +26,58 @@ import FreeCADGui as Gui
 
 from PySide2.QtWidgets import QDialogButtonBox
 
-import local
-from featureClasses import SlaveMasterGear, ViewProviderSlaveMasterGear
-from widgets import GearWidget
-from observers import SelObserver
+from freecad.invgears import local
+from freecad.invgears.featureClasses import InternalGear, SlaveGear, ViewProviderInternalGear, ViewProviderSlaveGear
+from freecad.invgears.widgets import GearWidget
 
 
-class createInvoluteGears3():
-    def __init__(self, fp_master, angle_list, widget):
+class createInvoluteGears4():
+    def __init__(self, widget):
 
-        list_s_gear = "[" + angle_list + "]"
+        body_internal = App.activeDocument().addObject('PartDesign::Body', 'body_internal')
+        fp_internal = App.activeDocument().addObject("PartDesign::FeaturePython", "fp_internal")
+        ViewProviderInternalGear(fp_internal.ViewObject)
+        InternalGear(fp_internal, body_internal, widget)
+        App.activeDocument().recompute()
+
+        list_s_gear = "[" + widget.list_angular_s_Edit.text() + "]"
         if list_s_gear:
             for item in eval(list_s_gear):
-                body_slave = App.activeDocument().addObject('PartDesign::Body', 'body_slave_master')
-                fp_slave = App.activeDocument().addObject("PartDesign::FeaturePython", "fp_slave_master")
-                ViewProviderSlaveMasterGear(fp_slave.ViewObject)
-                SlaveMasterGear(fp_slave, body_slave, fp_master, item, widget)
+                body_slave = App.activeDocument().addObject('PartDesign::Body', 'body_slave')
+                fp_slave = App.activeDocument().addObject("PartDesign::FeaturePython", "fp_slave")
+                ViewProviderSlaveGear(fp_slave.ViewObject)
+                SlaveGear(fp_slave, body_slave, fp_internal, item)
                 App.activeDocument().recompute()
 
         Gui.SendMsgToActiveView("ViewFit")
 
 
-class SlaveMasterGearTaskPanel:
+class InternalGearTaskPanel:
     def __init__(self):
-        self.form = GearWidget("SlaveMaster", self.attach_master_gear)
-        self.master_gear_Button = self.form.master_gear_Button
-        self.master_gear_Edit = self.form.master_gear_Edit
-        self.print_warnings = self.form.print_warnings
-        self.sel_o = SelObserver(self.master_gear_Button, self.master_gear_Edit, self.print_warnings)
-
-        self.attach_master_gear()
-
-    def attach_master_gear(self):
-        self.master_gear_Button.setText("Selecting master gear...")
-        self.master_gear_Button.setEnabled(False)
-        self.sel_o.addSelection(0, 0, 0, 0)
+        self.form = GearWidget("Internal")
+        print(type(self))
 
     def accept(self):
-        angle_list = self.form.list_angular_s_Edit2.text()
-        if not self.sel_o.sel_flag:
-            self.print_warnings.setText("First you have to choose a master gear and then click on ok.")
-        elif not self.form.list_angular_s_Edit2.text():
-            self.print_warnings.setText("First you have to put angular locations and then click on ok.")
+        N_m = self.form.N_m_Edit.text()
+        N_s = self.form.N_s_Edit.text()
+        if N_s >= N_m:
+            self.form.print_warnings.setText("Number of master gear teeth must be greater than Number of slave gear teeth")
         else:
-            createInvoluteGears3(self.sel_o.selection, angle_list, self.form)
-            Gui.Selection.removeObserver(self.sel_o)
+            createInvoluteGears4(self.form)
             Gui.Control.closeDialog()
 
     def reject(self):
-        Gui.Selection.removeObserver(self.sel_o)
         Gui.Control.closeDialog()
 
     def getStandardButtons(self):
         return QDialogButtonBox.Cancel | QDialogButtonBox.Ok
 
 
-class makeSlaveMasterGearCmd():
+class makeInternalGearCmd():
     def GetResources(self):
-        return {"MenuText": "Add Slave-Master Gear",
-                "ToolTip": "Add a new slave-master gear",
-                "Pixmap": local.path() + "/Resources/icons/slave-master_gear.svg"}
+        return {"MenuText": "Create Internal Gear",
+                "ToolTip": "Create a new internal gear",
+                "Pixmap": local.path() + "/Resources/icons/internal_gear.svg"}
 
     def IsActive(self):
         if App.activeDocument() is None:
@@ -94,8 +86,8 @@ class makeSlaveMasterGearCmd():
             return True
 
     def Activated(self):
-        panel = SlaveMasterGearTaskPanel()
+        panel = InternalGearTaskPanel()
         Gui.Control.showDialog(panel)
 
 
-Gui.addCommand('AddSlaveMasterGear', makeSlaveMasterGearCmd())
+Gui.addCommand('CreateInternalGear', makeInternalGearCmd())
