+++ freecad/invgears/newSVGCmd.py
@@ -26,10 +26,10 @@ import FreeCADGui as Gui
 
 from PySide2.QtWidgets import QDialogButtonBox, QFileDialog
 
-import local
-from svgFile import GearsInSVG
-from widgets import GearWidget
-from observers import SelObserver
+from freecad.invgears import local
+from freecad.invgears.svgFile import GearsInSVG
+from freecad.invgears.widgets import GearWidget
+from freecad.invgears.observers import SelObserver
 
 
 class createInvoluteGears5():
