+++ freecad/invgears/newSlaveMasterCmd.py
@@ -1,7 +1,7 @@
 # ***************************************************************************
 # *   Copyright (c) 2021 Sebastian Ernesto García <sebasg@outlook.com>      *
 # *                                                                         *
-# *   newMasterCmd.py                                                       *
+# *   newSlaveMasterCmd.py                                                  *
 # *                                                                         *
 # *   This program is free software; you can redistribute it and/or modify  *
 # *   it under the terms of the GNU Lesser General Public License (LGPL)    *
@@ -26,53 +26,66 @@ import FreeCADGui as Gui
 
 from PySide2.QtWidgets import QDialogButtonBox
 
-import local
-from featureClasses import MasterGear, SlaveGear, ViewProviderMasterGear, ViewProviderSlaveGear
-from widgets import GearWidget
+from freecad.invgears import local
+from freecad.invgears.featureClasses import SlaveMasterGear, ViewProviderSlaveMasterGear
+from freecad.invgears.widgets import GearWidget
+from freecad.invgears.observers import SelObserver
 
 
-class createInvoluteGears():
-    def __init__(self, widget):
+class createInvoluteGears3():
+    def __init__(self, fp_master, angle_list, widget):
 
-        body_master = App.activeDocument().addObject('PartDesign::Body', 'body_master')
-        fp_master = App.activeDocument().addObject("PartDesign::FeaturePython", "fp_master")
-        ViewProviderMasterGear(fp_master.ViewObject)
-        MasterGear(fp_master, body_master, widget)
-        App.activeDocument().recompute()
-
-        list_s_gear = "[" + widget.list_angular_s_Edit.text() + "]"
+        list_s_gear = "[" + angle_list + "]"
         if list_s_gear:
             for item in eval(list_s_gear):
-                body_slave = App.activeDocument().addObject('PartDesign::Body', 'body_slave')
-                fp_slave = App.activeDocument().addObject("PartDesign::FeaturePython", "fp_slave")
-                ViewProviderSlaveGear(fp_slave.ViewObject)
-                SlaveGear(fp_slave, body_slave, fp_master, item)
+                body_slave = App.activeDocument().addObject('PartDesign::Body', 'body_slave_master')
+                fp_slave = App.activeDocument().addObject("PartDesign::FeaturePython", "fp_slave_master")
+                ViewProviderSlaveMasterGear(fp_slave.ViewObject)
+                SlaveMasterGear(fp_slave, body_slave, fp_master, item, widget)
                 App.activeDocument().recompute()
 
         Gui.SendMsgToActiveView("ViewFit")
 
 
-class MasterGearTaskPanel:
+class SlaveMasterGearTaskPanel:
     def __init__(self):
-        self.form = GearWidget("Master")
-        print(type(self))
+        self.form = GearWidget("SlaveMaster", self.attach_master_gear)
+        self.master_gear_Button = self.form.master_gear_Button
+        self.master_gear_Edit = self.form.master_gear_Edit
+        self.print_warnings = self.form.print_warnings
+        self.sel_o = SelObserver(self.master_gear_Button, self.master_gear_Edit, self.print_warnings)
+
+        self.attach_master_gear()
+
+    def attach_master_gear(self):
+        self.master_gear_Button.setText("Selecting master gear...")
+        self.master_gear_Button.setEnabled(False)
+        self.sel_o.addSelection(0, 0, 0, 0)
 
     def accept(self):
-        createInvoluteGears(self.form)
-        Gui.Control.closeDialog()
+        angle_list = self.form.list_angular_s_Edit2.text()
+        if not self.sel_o.sel_flag:
+            self.print_warnings.setText("First you have to choose a master gear and then click on ok.")
+        elif not self.form.list_angular_s_Edit2.text():
+            self.print_warnings.setText("First you have to put angular locations and then click on ok.")
+        else:
+            createInvoluteGears3(self.sel_o.selection, angle_list, self.form)
+            Gui.Selection.removeObserver(self.sel_o)
+            Gui.Control.closeDialog()
 
     def reject(self):
+        Gui.Selection.removeObserver(self.sel_o)
         Gui.Control.closeDialog()
 
     def getStandardButtons(self):
         return QDialogButtonBox.Cancel | QDialogButtonBox.Ok
 
 
-class makeMasterGearCmd():
+class makeSlaveMasterGearCmd():
     def GetResources(self):
-        return {"MenuText": "Create Master Gear",
-                "ToolTip": "Create a new master gear",
-                "Pixmap": local.path() + "/Resources/icons/master_gear.svg"}
+        return {"MenuText": "Add Slave-Master Gear",
+                "ToolTip": "Add a new slave-master gear",
+                "Pixmap": local.path() + "/Resources/icons/slave-master_gear.svg"}
 
     def IsActive(self):
         if App.activeDocument() is None:
@@ -81,8 +94,8 @@ class makeMasterGearCmd():
             return True
 
     def Activated(self):
-        panel = MasterGearTaskPanel()
+        panel = SlaveMasterGearTaskPanel()
         Gui.Control.showDialog(panel)
 
 
-Gui.addCommand('CreateMasterGear', makeMasterGearCmd())
+Gui.addCommand('AddSlaveMasterGear', makeSlaveMasterGearCmd())
